refactor: move activity service to Activity package
