+++ src/main/java/com/alkemy/ong/web/controller/ActivityController.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.web.controller;
 
 import com.alkemy.ong.domain.model.Activity;
-import com.alkemy.ong.domain.service.ActivityService;
+import com.alkemy.ong.domain.activity.ActivityService;
 import com.alkemy.ong.web.dto.ActivityDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
